requirements.txt
Flask
textblob
google-generativeai
gunicorn